requirements.txt
pandas
sqlalchemy
PyYAML
psycopg2  # Conector para PostgreSQL
pymysql   # Conector para MySQL